pyproject.toml
[project]
name = "rag4tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "llama-index-core>=0.13.0",
    "llama-index-postprocessor-flag-embedding-reranker>=0.4.0",
    "numpy>=2.3.2",
    "openai>=1.98.0",
    "python-dotenv>=1.1.1",
    "redis>=6.2.0",
    "redisvl>=0.8.0",
    "requests>=2.32.4",
    "uuid>=1.30",
    "uvicorn>=0.35.0",
]
